refactor: split foundation helper types
